WIP: Refactor for multithreading
Making PieBreak inherit from Thread leaves the script
free to take new parameters while the timer is running. At
the moment it simply returns whatever is entered at the
terminal. It will be able to return information on how long
is left in the current loop and what urls are in use, as
well as add new urls to the list.