feat: Add a cli for scheduling, and make it downloadable
sched will not directly work with dynamic updates, and is
behaving strangely with threading. To my chagrin, this iteration
uses sched and doesn't attempt any advanced scheduling.

It is the first useful version of piebreak. Coming iterations
may add support for different playlist specification and
schedules with APScheduling, or maybe a QT5 GUI.

aside: Could be an opportunity to try using redis?